+++ src/main/java/com/microsoft/store/partnercenter/subscriptions/ISubscriptionActivationLinks.java
@@ -1,0 +1,22 @@
+// -----------------------------------------------------------------------
+// <copyright file="ISubscriptionActivationLinks.java" company="Microsoft">
+//      Copyright (c) Microsoft Corporation. All rights reserved.
+// </copyright>
+// -----------------------------------------------------------------------
+
+package com.microsoft.store.partnercenter.subscriptions;
+
+import com.microsoft.store.partnercenter.IPartnerComponent;
+import com.microsoft.store.partnercenter.genericoperations.IEntireEntityCollectionRetrievalOperations;
+import com.microsoft.store.partnercenter.models.ResourceCollection;
+import com.microsoft.store.partnercenter.models.orders.OrderLineItemActivationLink;
+import com.microsoft.store.partnercenter.models.utils.Tuple;
+
+/**
+ * A resource collection of subscription activation links.
+ */
+public interface ISubscriptionActivationLinks
+    extends IPartnerComponent<Tuple<String, String>>,
+    IEntireEntityCollectionRetrievalOperations<OrderLineItemActivationLink, ResourceCollection<OrderLineItemActivationLink>>
+{
+}